+++ src/marionette.collectionview.js
@@ -11,22 +11,22 @@ Marionette.CollectionView = Marionette.View.extend({
   // constructor
   constructor: function(options){
     this.initChildViewStorage();
+    this.onShowCallbacks = new Marionette.Callbacks();
 
     var args = Array.prototype.slice.apply(arguments);
     Marionette.View.prototype.constructor.apply(this, args);
 
-    this.initialEvents();
-    this.onShowCallbacks = new Marionette.Callbacks();
+    this._initialEvents();
   },
 
   // Configured the initial events that the collection view
   // binds to. Override this method to prevent the initial
   // events, or to add your own initial events.
-  initialEvents: function(){
+  _initialEvents: function(){
     if (this.collection){
-      this.bindTo(this.collection, "add", this.addChildView, this);
-      this.bindTo(this.collection, "remove", this.removeItemView, this);
-      this.bindTo(this.collection, "reset", this.render, this);
+      this.listenTo(this.collection, "add", this.addChildView, this);
+      this.listenTo(this.collection, "remove", this.removeItemView, this);
+      this.listenTo(this.collection, "reset", this.render, this);
     }
   },
 
@@ -151,6 +151,9 @@ Marionette.CollectionView = Marionette.View.extend({
     // set up the child view event forwarding
     this.addChildViewEventForwarding(view);
 
+    // this view is about to be added
+    this.triggerMethod("before:item:added", view);
+
     // Store the child view itself so we can properly
     // remove and/or close it later
     this.children.add(view);
@@ -158,13 +161,14 @@ Marionette.CollectionView = Marionette.View.extend({
     // Render it and show it
     var renderResult = this.renderItemView(view, index);
 
-    // let the world know this view was added
-    this.triggerMethod("item:added", view);
+    // this view was added
+    this.triggerMethod("after:item:added", view);
 
     // call onShow for child item views
-    if (view.onShow){
-      this.onShowCallbacks.add(view.onShow, view);
-    }
+    this.onShowCallbacks.add(function() {
+      this.triggerMethod.call(view, "show");
+    }, this);
+
 
     return renderResult;
   },
@@ -176,18 +180,13 @@ Marionette.CollectionView = Marionette.View.extend({
 
     // Forward all child item view events through the parent,
     // prepending "itemview:" to the event name
-    var childBinding = this.bindTo(view, "all", function(){
+    this.listenTo(view, "all", function(){
       var args = slice.call(arguments);
       args[0] = prefix + ":" + args[0];
       args.splice(1, 0, view);
 
-      this.triggerMethod.apply(this, args);
+      Marionette.triggerMethod.apply(this, args);
     }, this);
-
-    // Store all child event bindings so we can unbind
-    // them when removing / closing the child view
-    this._childBindings = this._childBindings || {};
-    this._childBindings[view.cid] = childBinding;
   },
 
   // render the item view
@@ -208,11 +207,7 @@ Marionette.CollectionView = Marionette.View.extend({
     var view = this.children.findByModel(item);
 
     if (view){
-      var childBinding = this._childBindings[view.cid];
-      if (childBinding) {
-        this.unbindFrom(childBinding);
-        delete this._childBindings[view.cid];
-      }
+      this.stopListening(view);
 
       if (view.close){
         view.close();